+++ src/native/http_connection_manager.c
@@ -234,7 +234,7 @@ JNIEXPORT jlong JNICALL Java_software_amazon_awssdk_crt_http_HttpClientConnectio
         &proxy_options,
         &proxy_tls_conn_options,
         jni_proxy_host,
-        jni_proxy_port,
+        (uint16_t)jni_proxy_port,
         jni_proxy_authorization_username,
         jni_proxy_authorization_password,
         jni_proxy_authorization_type,
